Data module restructuring (#13)
* split data module into datasetzoo and datautils #5

* moved dataset loading funcs to class files, changed attribute sanity check
