--- docs/source/api/neuralhydrology.data.caravan.rst
@@ -1,7 +1,0 @@
-Caravan
-=======
-
-.. automodule:: neuralhydrology.data.caravan
-   :members:
-   :undoc-members:
-   :show-inheritance:
