--- docs/source/api/neuralhydrology.data.pet.rst
@@ -1,7 +1,0 @@
-pet
-===
-
-.. automodule:: neuralhydrology.data.pet
-   :members:
-   :undoc-members:
-   :show-inheritance:
