--- docs/source/api/neuralhydrology.data.rst
@@ -1,20 +1,0 @@
-nh.data
-=======
-
-.. automodule:: neuralhydrology.data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. toctree::
-   :maxdepth: 4
-
-   neuralhydrology.data.basedataset
-   neuralhydrology.data.camelsus
-   neuralhydrology.data.hourlycamelsus
-   neuralhydrology.data.camelsgb
-   neuralhydrology.data.caravan
-   neuralhydrology.data.climateindices
-   neuralhydrology.data.dischargeinput
-   neuralhydrology.data.pet
-   neuralhydrology.data.utils
