+++ docs/source/api/neuralhydrology.datasetzoo.camelsus.rst
@@ -1,7 +1,7 @@
 CamelsUS
 ========
 
-.. automodule:: neuralhydrology.data.camelsus
+.. automodule:: neuralhydrology.datasetzoo.camelsus
    :members:
    :undoc-members:
    :show-inheritance:
