+++ docs/source/api/neuralhydrology.datasetzoo.hourlycamelsus.rst
@@ -1,7 +1,7 @@
 HourlyCamelsUS
 ==============
 
-.. automodule:: neuralhydrology.data.hourlycamelsus
+.. automodule:: neuralhydrology.datasetzoo.hourlycamelsus
    :members:
    :undoc-members:
    :show-inheritance:
