+++ docs/source/api/neuralhydrology.datasetzoo.rst
@@ -1,0 +1,15 @@
+nh.datasetzoo
+=============
+
+.. automodule:: neuralhydrology.datasetzoo
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. toctree::
+   :maxdepth: 4
+
+   neuralhydrology.datasetzoo.basedataset
+   neuralhydrology.datasetzoo.camelsus
+   neuralhydrology.datasetzoo.hourlycamelsus
+   neuralhydrology.datasetzoo.camelsgb
