+++ docs/source/api/neuralhydrology.datautils.climateindices.rst
@@ -1,7 +1,7 @@
 climateindices
 ==============
 
-.. automodule:: neuralhydrology.data.climateindices
+.. automodule:: neuralhydrology.datautils.climateindices
    :members:
    :undoc-members:
    :show-inheritance:
