+++ docs/source/api/neuralhydrology.datautils.dischargeinput.rst
@@ -1,7 +1,7 @@
 dischargeinput
 ==============
 
-.. automodule:: neuralhydrology.data.dischargeinput
+.. automodule:: neuralhydrology.datautils.dischargeinput
    :members:
    :undoc-members:
    :show-inheritance:
