+++ docs/source/api/neuralhydrology.datautils.rst
@@ -1,0 +1,15 @@
+nh.datautils
+============
+
+.. automodule:: neuralhydrology.datautils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. toctree::
+   :maxdepth: 4
+
+   neuralhydrology.datautils.climateindices
+   neuralhydrology.datautils.dischargeinput
+   neuralhydrology.datautils.pet
+   neuralhydrology.datautils.utils
