+++ docs/source/api/neuralhydrology.datautils.utils.rst
@@ -1,0 +1,7 @@
+utils
+=====
+
+.. automodule:: neuralhydrology.datautils.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
