--- neuralhydrology/data/camelsus.py
@@ -1,106 +1,0 @@
-from typing import Dict, List, Union
-
-import numpy as np
-import pandas as pd
-import xarray
-
-from neuralhydrology.data.basedataset import BaseDataset
-from neuralhydrology.data import utils
-from neuralhydrology.utils.config import Config
-
-
-class CamelsUS(BaseDataset):
-    """Data set class for the CAMELS US data set by [#]_ and [#]_.
-    
-    Parameters
-    ----------
-    cfg : Config
-        The run configuration.
-    is_train : bool 
-        Defines if the dataset is used for training or evaluating. If True (training), means/stds for each feature
-        are computed and stored to the run directory. If one-hot encoding is used, the mapping for the one-hot encoding 
-        is created and also stored to disk. If False, a `scaler` input is expected and similarly the `id_to_int` input
-        if one-hot encoding is used. 
-    period : {'train', 'validation', 'test'}
-        Defines the period for which the data will be loaded
-    basin : str, optional
-        If passed, the data for only this basin will be loaded. Otherwise the basin(s) are read from the appropriate
-        basin file, corresponding to the `period`.
-    additional_features : List[Dict[str, pd.DataFrame]], optional
-        List of dictionaries, mapping from a basin id to a pandas DataFrame. This DataFrame will be added to the data
-        loaded from the dataset and all columns are available as 'dynamic_inputs', 'static_inputs' and 
-        'target_variables'
-    id_to_int : Dict[str, int], optional
-        If the config argument 'use_basin_id_encoding' is True in the config and period is either 'validation' or 
-        'test', this input is required. It is a dictionary, mapping from basin id to an integer (the one-hot encoding).
-    scaler : Dict[str, Union[pd.Series, xarray.DataArray]], optional
-        If period is either 'validation' or 'test', this input is required. It contains the means and standard 
-        deviations for each feature and is stored to the run directory during training (train_data/train_data_scaler.p)
-        
-    References
-    ----------
-    .. [#] A. J. Newman, M. P. Clark, K. Sampson, A. Wood, L. E. Hay, A. Bock, R. J. Viger, D. Blodgett, 
-        L. Brekke, J. R. Arnold, T. Hopson, and Q. Duan: Development of a large-sample watershed-scale 
-        hydrometeorological dataset for the contiguous USA: dataset characteristics and assessment of regional 
-        variability in hydrologic model performance. Hydrol. Earth Syst. Sci., 19, 209-223, 
-        doi:10.5194/hess-19-209-2015, 2015
-    .. [#] Addor, N., Newman, A. J., Mizukami, N. and Clark, M. P.: The CAMELS data set: catchment attributes and 
-        meteorology for large-sample studies, Hydrol. Earth Syst. Sci., 21, 5293-5313, doi:10.5194/hess-21-5293-2017,
-        2017.
-    """
-
-    def __init__(self,
-                 cfg: Config,
-                 is_train: bool,
-                 period: str,
-                 basin: str = None,
-                 additional_features: List[Dict[str, pd.DataFrame]] = [],
-                 id_to_int: Dict[str, int] = {},
-                 scaler: Dict[str, Union[pd.Series, xarray.DataArray]] = {}):
-        super(CamelsUS, self).__init__(cfg=cfg,
-                                       is_train=is_train,
-                                       period=period,
-                                       basin=basin,
-                                       additional_features=additional_features,
-                                       id_to_int=id_to_int,
-                                       scaler=scaler)
-
-    def _load_basin_data(self, basin: str) -> pd.DataFrame:
-        """Load input and output data from text files."""
-        # get forcings
-        dfs = []
-        for forcing in self.cfg.forcings:
-            df, area = utils.load_camels_us_forcings(self.cfg.data_dir, basin, forcing)
-
-            # rename columns
-            if len(self.cfg.forcings) > 1:
-                df = df.rename(columns={col: f"{col}_{forcing}" for col in df.columns})
-            dfs.append(df)
-        df = pd.concat(dfs, axis=1)
-
-        # add discharge
-        df['QObs(mm/d)'] = utils.load_camels_us_discharge(self.cfg.data_dir, basin, area)
-
-        # replace invalid discharge values by NaNs
-        qobs_cols = [col for col in df.columns if "qobs" in col.lower()]
-        for col in qobs_cols:
-            df.loc[df[col] < 0, col] = np.nan
-
-        return df
-
-    def _load_attributes(self) -> pd.DataFrame:
-        if self.cfg.camels_attributes:
-            if self.is_train:
-                # sanity check attributes for NaN in per-feature standard deviation
-                utils.attributes_sanity_check(data_dir=self.cfg.data_dir,
-                                              attribute_set=self.cfg.dataset,
-                                              basins=self.basins,
-                                              attribute_list=self.cfg.camels_attributes)
-
-            df = utils.load_camels_us_attributes(self.cfg.data_dir, basins=self.basins)
-
-            # remove all attributes not defined in the config
-            drop_cols = [c for c in df.columns if c not in self.cfg.camels_attributes]
-            df = df.drop(drop_cols, axis=1)
-
-            return df
