+++ neuralhydrology/datasetzoo/hourlycamelsus.py
@@ -1,10 +1,12 @@
 import logging
 import pickle
+from pathlib import Path
 
 import numpy as np
 import pandas as pd
+import xarray
 
-from neuralhydrology.data import utils, CamelsUS
+from neuralhydrology.datasetzoo.camelsus import CamelsUS, load_camels_us_forcings, load_camels_us_attributes
 from neuralhydrology.utils.config import Config
 
 LOGGER = logging.getLogger(__name__)
@@ -71,7 +73,7 @@ class HourlyCamelsUS(CamelsUS):
                 df = self.load_hourly_data(basin, forcing)
             else:
                 # load daily CAMELS forcings and upsample to hourly
-                df, _ = utils.load_camels_us_forcings(self.cfg.data_dir, basin, forcing)
+                df, _ = load_camels_us_forcings(self.cfg.data_dir, basin, forcing)
                 df = df.resample('1H').ffill()
             if len(self.cfg.forcings) > 1:
                 # rename columns
@@ -86,12 +88,12 @@ class HourlyCamelsUS(CamelsUS):
 
         # add stage, if requested
         if 'gauge_height_m' in self.cfg.target_variables:
-            df = df.join(utils.load_hourly_us_stage(self.cfg.data_dir, basin))
+            df = df.join(load_hourly_us_stage(self.cfg.data_dir, basin))
             df.loc[df['gauge_height_m'] < 0, 'gauge_height_m'] = np.nan
 
         # convert discharge to 'synthetic' stage, if requested
         if 'synthetic_qobs_stage_meters' in self.cfg.target_variables:
-            attributes = utils.load_camels_us_attributes(data_dir=self.cfg.data_dir, basins=[basin])
+            attributes = load_camels_us_attributes(data_dir=self.cfg.data_dir, basins=[basin])
             with open(self.cfg.rating_curve_file, 'rb') as f:
                 rating_curves = pickle.load(f)
             df['synthetic_qobs_stage_meters'] = np.nan
@@ -119,7 +121,7 @@ class HourlyCamelsUS(CamelsUS):
         fallback_csv = False
         try:
             if self._netcdf_dataset is None:
-                self._netcdf_dataset = utils.load_hourly_us_netcdf(self.cfg.data_dir, forcings)
+                self._netcdf_dataset = load_hourly_us_netcdf(self.cfg.data_dir, forcings)
             df = self._netcdf_dataset.sel(basin=basin).to_dataframe()
         except FileNotFoundError:
             fallback_csv = True
@@ -130,9 +132,130 @@ class HourlyCamelsUS(CamelsUS):
             fallback_csv = True
             LOGGER.warning(f'## Warning: NetCDF file does not contain data for {basin}. Trying slower csv files.')
         if fallback_csv:
-            df = utils.load_hourly_us_forcings(self.cfg.data_dir, basin, forcings)
+            df = load_hourly_us_forcings(self.cfg.data_dir, basin, forcings)
 
             # add discharge
-            df = df.join(utils.load_hourly_us_discharge(self.cfg.data_dir, basin))
+            df = df.join(load_hourly_us_discharge(self.cfg.data_dir, basin))
 
         return df
+
+
+def load_hourly_us_forcings(data_dir: Path, basin: str, forcings: str) -> pd.DataFrame:
+    """Load the hourly forcing data for a basin of the CAMELS US data set.
+
+    The hourly forcings are not included in the original data set by Newman et al. (2017).
+
+    Parameters
+    ----------
+    data_dir : Path
+        Path to the CAMELS US directory. This folder must contain an 'hourly' folder containing one subdirectory
+        for each forcing, which contains the forcing files (.csv) for each basin. Files have to contain the 8-digit 
+        basin id.
+    basin : str
+        8-digit USGS identifier of the basin.
+    forcings : str
+        Must match the folder names in the 'hourly' directory. E.g. 'nldas_hourly'
+
+    Returns
+    -------
+    pd.DataFrame
+        Time-indexed DataFrame, containing the forcing data.
+    """
+    forcing_path = data_dir / 'hourly' / forcings
+    if not forcing_path.is_dir():
+        raise OSError(f"{forcing_path} does not exist")
+
+    files = list(forcing_path.glob('*.csv'))
+    file_path = [f for f in files if basin in f.stem]
+    if file_path:
+        file_path = file_path[0]
+    else:
+        raise FileNotFoundError(f'No file for Basin {basin} at {forcing_path}')
+
+    return pd.read_csv(file_path, index_col=['date'], parse_dates=['date'])
+
+
+def load_hourly_us_discharge(data_dir: Path, basin: str) -> pd.DataFrame:
+    """Load the hourly discharge data for a basin of the CAMELS US data set.
+
+    Parameters
+    ----------
+    data_dir : Path
+        Path to the CAMELS US directory. This folder must contain a folder called 'hourly' with a subdirectory 
+        'usgs_streamflow' which contains the discharge files (.csv) for each basin. File names must contain the 8-digit 
+        basin id.
+    basin : str
+        8-digit USGS identifier of the basin.
+
+    Returns
+    -------
+    pd.Series
+        Time-index Series of the discharge values (mm/hour)
+    """
+    discharge_path = data_dir / 'hourly' / 'usgs_streamflow'
+    files = list(discharge_path.glob('**/*usgs-hourly.csv'))
+    file_path = [f for f in files if basin in f.stem]
+    if file_path:
+        file_path = file_path[0]
+    else:
+        raise FileNotFoundError(f'No file for Basin {basin} at {discharge_path}')
+
+    return pd.read_csv(file_path, index_col=['date'], parse_dates=['date'])
+
+
+def load_hourly_us_stage(data_dir: Path, basin: str) -> pd.Series:
+    """Load the hourly stage data for a basin of the CAMELS US data set.
+
+    Parameters
+    ----------
+    data_dir : Path
+        Path to the CAMELS US directory. This folder must contain a folder called 'hourly' with a subdirectory 
+        'usgs_stage' which contains the stage files (.csv) for each basin. File names must contain the 8-digit basin id.
+    basin : str
+        8-digit USGS identifier of the basin.
+
+    Returns
+    -------
+    pd.Series
+        Time-index Series of the stage values (m)
+    """
+    stage_path = data_dir / 'hourly' / 'usgs_stage'
+    files = list(stage_path.glob('**/*_utc.csv'))
+    file_path = [f for f in files if basin in f.stem]
+    if file_path:
+        file_path = file_path[0]
+    else:
+        raise FileNotFoundError(f'No file for Basin {basin} at {stage_path}')
+
+    df = pd.read_csv(file_path,
+                     sep=',',
+                     index_col=['datetime'],
+                     parse_dates=['datetime'],
+                     usecols=['datetime', 'gauge_height_ft'])
+    df = df.resample('H').mean()
+    df["gauge_height_m"] = df["gauge_height_ft"] * 0.3048
+
+    return df["gauge_height_m"]
+
+
+def load_hourly_us_netcdf(data_dir: Path, forcings: str) -> xarray.Dataset:
+    """Load hourly forcing and discharge data from preprocessed netCDF file.
+
+    Parameters
+    ----------
+    data_dir : Path
+        Path to the CAMELS US directory. This folder must contain a folder called 'hourly', containing the netCDF file.
+    forcings : str
+        Name of the forcing product. Must match the ending of the netCDF file. E.g. 'nldas_hourly' for 
+        'usgs-streamflow-nldas_hourly.nc'
+
+    Returns
+    -------
+    xarray.Dataset
+        Dataset containing the combined discharge and forcing data of all basins (as stored in the netCDF)  
+    """
+    netcdf_path = data_dir / 'hourly' / f'usgs-streamflow-{forcings}.nc'
+    if not netcdf_path.is_file():
+        raise FileNotFoundError(f'No NetCDF file for hourly streamflow and {forcings} at {netcdf_path}.')
+
+    return xarray.open_dataset(netcdf_path)
