+++ neuralhydrology/modelzoo/mtslstm.py
@@ -5,7 +5,7 @@ import pandas as pd
 import torch
 import torch.nn as nn
 
-from neuralhydrology.data.utils import sort_frequencies
+from neuralhydrology.datautils.utils import sort_frequencies
 from neuralhydrology.modelzoo.head import get_head
 from neuralhydrology.modelzoo.basemodel import BaseModel
 from neuralhydrology.utils.config import Config
